Refactor release matching logic into util class; move tests
