+++ src/test/java/uk/gov/companieshouse/servicesdashboardapi/model/github/GitInfoTest.java
@@ -6,7 +6,6 @@ import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.junit.jupiter.api.Assertions.assertSame;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 class GitInfoTest {
@@ -44,6 +43,7 @@ class GitInfoTest {
         assertEquals(1, gitInfo.getReleases().size());
         assertEquals("1.0.0", gitInfo.getReleases().get(0).getVersion());
         assertEquals("2026-01-01", gitInfo.getReleases().get(0).getDate());
+        assertEquals("{r:null l:null o:null sA:null [R:[{v:1.0.0,d:2026-01-01}]]}", gitInfo.toString());
     }
 
     @Test
@@ -54,112 +54,5 @@ class GitInfoTest {
 
         assertNotNull(gitInfo.getReleases());
         assertTrue(gitInfo.getReleases().isEmpty());
-    }
-
-    @Test
-    void shouldKeepFirstReleaseAndFirstDifferentReleaseCycle() {
-        GitInfo gitInfo = new GitInfo();
-
-        GitReleaseInfo first = new GitReleaseInfo();
-        first.setVersion("ecs-service-1.0.22");
-        first.setDate("2026-01-22");
-
-        GitReleaseInfo sameCycleOne = new GitReleaseInfo();
-        sameCycleOne.setVersion("ecs-service-1.0.21");
-        sameCycleOne.setDate("2026-01-21");
-
-        GitReleaseInfo sameCycleTwo = new GitReleaseInfo();
-        sameCycleTwo.setVersion("ecs-service-1.0.20");
-        sameCycleTwo.setDate("2026-01-20");
-
-        GitReleaseInfo differentCycle = new GitReleaseInfo();
-        differentCycle.setVersion("4.0.11");
-        differentCycle.setDate("2025-12-11");
-
-        List<GitReleaseInfo> releases = new ArrayList<>();
-        releases.add(first);
-        releases.add(sameCycleOne);
-        releases.add(sameCycleTwo);
-        releases.add(differentCycle);
-
-        gitInfo.setReleases(releases);
-
-        assertNotNull(gitInfo.getReleases());
-        assertEquals(2, gitInfo.getReleases().size());
-        assertEquals("ecs-service-1.0.22", gitInfo.getReleases().get(0).getVersion());
-        assertEquals("4.0.11", gitInfo.getReleases().get(1).getVersion());
-    }
-
-    @Test
-    void shouldAddFirstNonMatchingReleaseToFilteredList() {
-        GitInfo gitInfo = new GitInfo();
-
-        GitReleaseInfo first = new GitReleaseInfo();
-        first.setVersion("1.2.3");
-        first.setDate("2026-01-23");
-
-        GitReleaseInfo nonMatching = new GitReleaseInfo();
-        nonMatching.setVersion("release-1.2.3");
-        nonMatching.setDate("2026-01-22");
-
-        List<GitReleaseInfo> releases = new ArrayList<>();
-        releases.add(first);
-        releases.add(nonMatching);
-
-        gitInfo.setReleases(releases);
-
-        assertEquals(2, gitInfo.getReleases().size());
-        assertSame(first, gitInfo.getReleases().get(0));
-        assertSame(nonMatching, gitInfo.getReleases().get(1));
-    }
-
-    @Test
-    void shouldOnlyAddOneIfAllMatchingVersions() {
-        GitInfo gitInfo = new GitInfo();
-
-        GitReleaseInfo first = new GitReleaseInfo();
-        first.setVersion("1.2.3");
-        first.setDate("2026-01-23");
-
-        GitReleaseInfo matching = new GitReleaseInfo();
-        matching.setVersion("1.2.3");
-        matching.setDate("2026-01-22");
-
-        List<GitReleaseInfo> releases = new ArrayList<>();
-        releases.add(first);
-        releases.add(matching);
-
-        gitInfo.setReleases(releases);
-
-        assertEquals(1, gitInfo.getReleases().size());
-        assertSame(first, gitInfo.getReleases().get(0));
-    }
-
-    @Test
-    void shouldHandleAlphabeticReleasePatternWhenFiltering() {
-        GitInfo gitInfo = new GitInfo();
-
-        GitReleaseInfo first = new GitReleaseInfo();
-        first.setVersion("abc");
-        first.setDate("2026-01-23");
-
-        GitReleaseInfo matchingAlphabetic = new GitReleaseInfo();
-        matchingAlphabetic.setVersion("xyz");
-        matchingAlphabetic.setDate("2026-01-22");
-
-        GitReleaseInfo differentPattern = new GitReleaseInfo();
-        differentPattern.setVersion("123");
-        differentPattern.setDate("2026-01-21");
-
-        List<GitReleaseInfo> releases = new ArrayList<>();
-        releases.add(first);
-        releases.add(matchingAlphabetic);
-        releases.add(differentPattern);
-
-        gitInfo.setReleases(releases);
-
-        assertEquals(2, gitInfo.getReleases().size());
-        assertSame(first, gitInfo.getReleases().get(0));
-        assertSame(differentPattern, gitInfo.getReleases().get(1));
     }
 }
